Refactor the run_time ============================= - Moved from uneffective windows_overlaped to native_workers_singleton - Removed unused librarys: enbt, mem_tool, saloc, string_convert - Refactored compile time configuration,     now everything is in configuration/tasks.hpp - New Tasks feature: callback tasks - New STD library: bytes, used for switching between endianess - Implemented STD library: file - Refactored STD library: net     now Implemented tcp client and udp client/server - Console reads now more safe by checking if the input is valid utf8 - Progress: seh exceptions implementing step by step(destructors work) - STD documentation is now completed and can be found in docks folder

+++ run_time/attacha_abi.hpp
@@ -219,7 +219,7 @@ namespace ABI_IMPL {
 				else if constexpr (std::is_same_v<T, list_array<ValueItem>>)
 					return { ValueItem(val,meta) };
 				else if constexpr (std::is_arithmetic_v<T>)
-					return (T)(int8_t)val;
+					return (T)(int8_t)(ptrdiff_t)val;
 				else if constexpr (std::is_arithmetic_v<std::remove_pointer_t<T>>)
 					return AsPointer<T, int8_t>(val);
 				else
@@ -232,7 +232,7 @@ namespace ABI_IMPL {
 				else if constexpr (std::is_same_v<T, list_array<ValueItem>>)
 					return { ValueItem(val,meta) };
 				else if constexpr (std::is_arithmetic_v<T>)
-					return (T)(int16_t)val;
+					return (T)(int16_t)(ptrdiff_t)val;
 				else if constexpr (std::is_arithmetic_v<std::remove_pointer_t<T>>)
 					return AsPointer<T, int16_t>(val);
 				else
@@ -245,7 +245,7 @@ namespace ABI_IMPL {
 				else if constexpr (std::is_same_v<T, list_array<ValueItem>>)
 					return { ValueItem(val,meta) };
 				else if constexpr (std::is_arithmetic_v<T>)
-					return (T)(int32_t)val;
+					return (T)(int32_t)(ptrdiff_t)val;
 				else if constexpr (std::is_arithmetic_v<std::remove_pointer_t<T>>)
 					return AsPointer<T, int32_t>(val);
 				else
@@ -258,7 +258,7 @@ namespace ABI_IMPL {
 				else if constexpr (std::is_same_v<T, list_array<ValueItem>>)
 					return { ValueItem(val,meta) };
 				else if constexpr (std::is_arithmetic_v<T>)
-					return (T)(int64_t)val;
+					return (T)(int64_t)(ptrdiff_t)val;
 				else if constexpr (std::is_arithmetic_v<std::remove_pointer_t<T>>)
 					return AsPointer<T, int64_t>(val);
 				else
